+++ changelog.md
@@ -1,4 +1,5 @@
 ##### Version 0.16-109
+- Requires Forge version 32.0.67 or later (BluSunrize)
 - Fixed recipes not loading on dedicated servers (BluSunrize)
 - Fixed fluid tank backgrounds being wrongly offset in JEI (BluSunrize)
 - Fixed revolver causing a crash with its attribute map (BluSunrize)